refactor redis fns to separate file
