+++ src/main.rs
@@ -3,6 +3,7 @@ mod rpc_conf;
 mod shared_storage;
 mod signing;
 mod rpc;
+mod redis_fns;
 
 use axum::{
     extract::Json,
@@ -26,11 +27,9 @@ use near_primitives::transaction::Action;
 #[cfg(test)]
 use near_primitives::types::{BlockHeight, Nonce};
 use near_primitives::types::AccountId;
-use near_primitives::views::ExecutionOutcomeWithIdView;
 use once_cell::sync::Lazy;
-use r2d2::{Pool, PooledConnection};
-use r2d2_redis::{redis, RedisConnectionManager};
-use r2d2_redis::redis::{Commands, RedisError};
+use r2d2::Pool;
+use r2d2_redis::RedisConnectionManager;
 use serde::Deserialize;
 use serde_json::json;
 use std::fmt;
@@ -328,7 +327,7 @@ async fn get_allowance(account_id_json: Json<AccountIdJson>) -> impl IntoRespons
     let Ok(account_id_val) = AccountId::from_str(&account_id_json.account_id) else {
         return (StatusCode::FORBIDDEN, format!("Invalid account_id: {}", account_id_json.account_id)).into_response();
     };
-    match get_remaining_allowance(&account_id_val).await {
+    match redis_fns::get_remaining_allowance(&account_id_val).await {
         Ok(allowance) => (
             StatusCode::OK,
             allowance.to_string()  // TODO: LP return in json format
@@ -345,33 +344,6 @@ async fn get_allowance(account_id_json: Json<AccountIdJson>) -> impl IntoRespons
             (StatusCode::INTERNAL_SERVER_ERROR, err_msg).into_response()
         }
     }
-}
-
-async fn get_redis_cnxn() -> Result<PooledConnection<RedisConnectionManager>, RedisError> {
-    let conn_result = REDIS_POOL.get();
-    let conn: PooledConnection<RedisConnectionManager> = match conn_result {
-        Ok(conn) => conn,
-        Err(e) => {
-            return Err(redis::RedisError::from((
-                redis::ErrorKind::IoError,
-                "Error getting Relayer DB connection from the pool",
-                e.to_string(),
-            )));
-        }
-    };
-    Ok(conn)
-}
-
-async fn set_account_and_allowance_in_redis(
-    account_id: &str,
-    allowance_in_gas: &u64,
-) -> Result<(), RedisError> {
-    // Get a connection from the REDIS_POOL
-    let mut conn = get_redis_cnxn().await?;
-
-    // Save the allowance information to Redis
-    conn.set(account_id, allowance_in_gas.clone())?;
-    Ok(())
 }
 
 
@@ -418,7 +390,7 @@ async fn create_account_atomic(
      */
 
     // check if the oauth_token has already been used and is a key in Redis
-    match get_oauth_token_in_redis(&oauth_token).await {
+    match redis_fns::get_oauth_token_in_redis(&oauth_token).await {
         Ok(is_oauth_token_in_redis) => {
             if is_oauth_token_in_redis {
                 let err_msg = format!(
@@ -437,7 +409,7 @@ async fn create_account_atomic(
             return (StatusCode::INTERNAL_SERVER_ERROR, err_msg).into_response();
         }
     }
-    let redis_result = set_account_and_allowance_in_redis(
+    let redis_result = redis_fns::set_account_and_allowance_in_redis(
         account_id,
         allowance_in_gas,
     ).await;
@@ -474,7 +446,7 @@ async fn create_account_atomic(
     }
 
     // add oauth token to redis (key: oauth_token, val: true)
-    match set_oauth_token_in_redis(oauth_token.clone()).await {
+    match redis_fns::set_oauth_token_in_redis(oauth_token.clone()).await {
         Ok(_) => {
             (
                 StatusCode::CREATED,
@@ -494,30 +466,6 @@ async fn create_account_atomic(
     }
 }
 
-async fn get_oauth_token_in_redis(
-    oauth_token: &str,
-) -> Result<bool, RedisError> {
-    // Get a connection from the REDIS_POOL
-    let mut conn = get_redis_cnxn().await?;
-    let is_already_used_option: Option<bool> = conn.get(oauth_token.to_owned())?;
-
-    match is_already_used_option {
-        Some(_) => Ok(true),
-        None => Ok(false),
-    }
-}
-
-async fn set_oauth_token_in_redis(
-    oauth_token: String,
-) -> Result<(), RedisError> {
-    // Get a connection from the REDIS_POOL
-    let mut conn = get_redis_cnxn().await?;
-
-    // Save the allowance information to Relayer DB
-    conn.set(&oauth_token, true)?;
-    Ok(())
-}
-
 
 #[utoipa::path(
 post,
@@ -535,7 +483,7 @@ async fn update_allowance(
     let account_id: &String = &account_id_allowance.account_id;
     let allowance_in_gas: &u64 = &account_id_allowance.allowance;
 
-    let redis_result = set_account_and_allowance_in_redis(
+    let redis_result = redis_fns::set_account_and_allowance_in_redis(
         account_id,
         allowance_in_gas,
     ).await;
@@ -556,61 +504,6 @@ async fn update_allowance(
     ).into_response()
 }
 
-async fn update_all_allowances_in_redis(allowance_in_gas: u64) -> Result<String, RelayError> {
-    // Get a connection to Redis from the pool
-    let mut redis_conn = match get_redis_cnxn().await {
-        Ok(conn) => conn,
-        Err(e) => {
-            let err_msg = format!("Error getting Relayer DB connection from the pool: {}", e);
-            error!("{err_msg}");
-            return Err(RelayError {
-                status_code: StatusCode::INTERNAL_SERVER_ERROR,
-                message: err_msg
-            });
-        }
-    };
-
-    // Fetch all keys that match the network env (.near for mainnet, .testnet for testnet, etc)
-    let network: String = if NETWORK_ENV.clone() != "mainnet" {
-        NETWORK_ENV.clone()
-    } else {
-        "near".to_string()
-    };
-    let pattern = format!("*.{}", network);
-    let keys: Vec<String> = match redis_conn.keys(pattern) {
-        Ok(keys) => keys,
-        Err(e) => {
-            let err_msg = format!("Error fetching keys from Relayer DB: {}", e);
-            error!("{err_msg}");
-            return Err(RelayError {
-                status_code: StatusCode::INTERNAL_SERVER_ERROR,
-                message: err_msg
-            });
-        }
-    };
-
-    // Iterate through the keys and update their values to the provided allowance in gas
-    for key in &keys {
-        match redis_conn.set::<_, _, ()>(key, allowance_in_gas.to_string()) {
-            Ok(_) => debug!("Updated allowance for key {}", key),
-            Err(e) => {
-                let err_msg = format!("Error updating allowance for key {}: {}", key, e);
-                error!("{err_msg}");
-                return Err(RelayError {
-                    status_code: StatusCode::INTERNAL_SERVER_ERROR,
-                    message: err_msg
-                });
-            }
-        }
-    }
-
-    // Return a success response
-    let num_keys = keys.len();
-    let success_msg: String = format!("Updated {num_keys:?} keys in Relayer DB");
-    info!("{success_msg}");
-    Ok(success_msg)
-}
-
 
 #[utoipa::path(
     post,
@@ -625,7 +518,7 @@ async fn update_all_allowances(
     Json(allowance_json): Json<AllowanceJson>,
 ) -> impl IntoResponse {
     let allowance_in_gas = allowance_json.allowance_in_gas;
-    let redis_response = update_all_allowances_in_redis(allowance_in_gas).await;
+    let redis_response = redis_fns::update_all_allowances_in_redis(allowance_in_gas).await;
     match redis_response {
         Ok(response) => response.into_response(),
         Err(err) => (err.status_code, err.message).into_response(),
@@ -650,7 +543,7 @@ async fn register_account_and_allowance(
     let allowance_in_gas: &u64 = &account_id_allowance_oauth.allowance;
     let oauth_token: &String = &account_id_allowance_oauth.oauth_token;
     // check if the oauth_token has already been used and is a key in Relayer DB
-    match get_oauth_token_in_redis(oauth_token).await {
+    match redis_fns::get_oauth_token_in_redis(oauth_token).await {
         Ok(is_oauth_token_in_redis) => {
             if is_oauth_token_in_redis {
                 let err_msg = format!(
@@ -670,7 +563,7 @@ async fn register_account_and_allowance(
             return (StatusCode::INTERNAL_SERVER_ERROR, err_msg).into_response();
         }
     }
-    let redis_result = set_account_and_allowance_in_redis(
+    let redis_result = redis_fns::set_account_and_allowance_in_redis(
         account_id,
         &allowance_in_gas,
     ).await;
@@ -693,7 +586,7 @@ async fn register_account_and_allowance(
     }
 
     // add oauth token to redis (key: oauth_token, val: true)
-    match set_oauth_token_in_redis(oauth_token.clone()).await {
+    match redis_fns::set_oauth_token_in_redis(oauth_token.clone()).await {
         Ok(_) => {
             (
                 StatusCode::CREATED,
@@ -708,33 +601,6 @@ async fn register_account_and_allowance(
             (StatusCode::INTERNAL_SERVER_ERROR, err_msg).into_response()
         }
     }
-}
-
-async fn get_remaining_allowance(
-    account_id: &AccountId,
-) -> Result<u64, RedisError> {
-    // Destructure the Extension and get a connection from the connection manager
-    let mut conn = get_redis_cnxn().await?;
-    let allowance: Option<u64> = conn.get(account_id.as_str())?;
-    let Some(remaining_allowance) = allowance else {
-        return Ok(0);
-    };
-    info!("get remaining allowance for account: {account_id}, {remaining_allowance}");
-    Ok(remaining_allowance)
-}
-
-// fn to update allowance in redis when getting the receipts back and deduct the gas used
-async fn update_remaining_allowance(
-    account_id: &AccountId,
-    gas_used_in_yn: u128,
-    allowance: u64,
-) -> Result<u64, RedisError> {
-    let mut conn = get_redis_cnxn().await?;
-    let key = account_id.clone().to_string();
-    let gas_used: u64 = (gas_used_in_yn / YN_TO_GAS) as u64;
-    let remaining_allowance = allowance - gas_used;
-    conn.set(key, remaining_allowance)?;
-    Ok(remaining_allowance.clone())
 }
 
 #[utoipa::path(
@@ -791,24 +657,6 @@ async fn send_meta_tx(
     }
 }
 
-fn calculate_total_gas_burned(
-    transaction_outcome: ExecutionOutcomeWithIdView,
-    execution_outcome: Vec<ExecutionOutcomeWithIdView>
-) -> u128 {
-    let mut total_tokens_burnt_in_yn: u128 = 0;
-    total_tokens_burnt_in_yn += transaction_outcome.outcome.tokens_burnt;
-
-    let exec_outcome_sum: u128 = execution_outcome
-        .iter()
-        .map(|ro| {
-            &ro.outcome.tokens_burnt
-        })
-        .sum();
-    total_tokens_burnt_in_yn += exec_outcome_sum;
-
-    total_tokens_burnt_in_yn
-}
-
 async fn process_signed_delegate_action(
     signed_delegate_action: SignedDelegateAction,
 ) -> Result<String, RelayError> {
@@ -858,7 +706,7 @@ async fn process_signed_delegate_action(
     if USE_REDIS.clone() {
         // Check the sender's remaining gas allowance in Redis
         let end_user_account: &AccountId = &signed_delegate_action.delegate_action.sender_id;
-        let remaining_allowance: u64 = get_remaining_allowance(end_user_account).await.unwrap_or(0);
+        let remaining_allowance: u64 = redis_fns::get_remaining_allowance(end_user_account).await.unwrap_or(0);
         if remaining_allowance < TXN_GAS_ALLOWANCE {
             let err_msg = format!(
                 "AccountId {} does not have enough remaining gas allowance.",
@@ -914,12 +762,12 @@ async fn process_signed_delegate_action(
             "Receipts Outcome": &execution.receipts_outcome,
         });
 
-        let gas_used_in_yn = calculate_total_gas_burned(
+        let gas_used_in_yn = redis_fns::calculate_total_gas_burned(
             execution.transaction_outcome,
             execution.receipts_outcome,
         );
         debug!("total gas burnt in yN: {}", gas_used_in_yn);
-        let new_allowance = update_remaining_allowance(
+        let new_allowance = redis_fns::update_remaining_allowance(
             &signer_account_id,
             gas_used_in_yn,
             remaining_allowance
@@ -1064,7 +912,7 @@ async fn test_send_meta_tx() {   // tests assume testnet in config
 
     // simulate calling the '/update_allowance' function with sender_id & allowance
     let allowance_in_gas: u64 = u64::MAX;
-    set_account_and_allowance_in_redis(&sender_id, &allowance_in_gas).await.expect(
+    redis_fns::set_account_and_allowance_in_redis(&sender_id, &allowance_in_gas).await.expect(
         "Failed to update account and allowance in redis"
     );
